chore: refactor utils functions to remove index.ts
