+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import './globals.css';
 import { lpmq, poppins } from '~/lib/fonts';
 import { NavigationBar } from '~/components/molecules';
 import { ThemeProvider } from '~/components/organisms';
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 export default function RootLayout({
   children,
