+++ src/components/atoms/button/button.component.tsx
@@ -1,7 +1,7 @@
 import { forwardRef } from 'react';
 import { Slot } from '@radix-ui/react-slot';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 import { buttonVariants } from './button.styles';
 
 import type { VariantProps } from 'class-variance-authority';
