+++ src/components/atoms/card/card.component.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from 'react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 const Card = forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
   ({ className, ...props }, ref) => (
