+++ src/components/atoms/dialog/dialog.component.tsx
@@ -13,7 +13,7 @@ import {
 } from '@radix-ui/react-dialog';
 import { X } from 'lucide-react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 const Dialog = Root;
