+++ src/components/atoms/dropdown-menu/dropdown-menu.component.tsx
@@ -11,7 +11,7 @@ import {
 } from '@radix-ui/react-dropdown-menu';
 import { forwardRef } from 'react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 const DropdownMenu = Root;
