+++ src/components/atoms/select/select.styles.ts
@@ -1,4 +1,4 @@
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 import type { ClassNamesConfig, GroupBase } from 'react-select';
