+++ src/components/atoms/skeleton/skeleton.component.tsx
@@ -1,4 +1,4 @@
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 const Skeleton = ({
   className,
