+++ src/components/atoms/table/table.component.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from 'react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 import TableRow from './table.client.component';
 
 const Table = forwardRef<
