+++ src/components/molecules/breadcrumb/breadcrumb.component.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 
 import { Separator } from '~/components/atoms';
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 export type BreadcrumbProps = React.HTMLAttributes<HTMLElement> & {
   paths: { title: string; link?: string }[];
