+++ src/components/molecules/search-button/search-button.component.tsx
@@ -2,7 +2,7 @@ import { Search } from 'lucide-react';
 
 import { SearchForm } from '../search-form';
 import { Button, Dialog, Separator } from '~/components/atoms';
-import { cn } from '~/lib/utils';
+import { cn } from '~/lib/utils/cn';
 
 import type { FC } from 'react';
 import type { DialogProps, DialogTriggerProps } from '@radix-ui/react-dialog';
