--- src/lib/utils/index.ts
@@ -1,22 +1,0 @@
-import { cn } from './cn';
-import { getBooks } from './getBooks';
-import { getCategories } from './categories.server';
-import { getContents } from './getContents';
-import { getInvoices } from './getInvoices';
-import { getPaths } from './getPaths';
-import { getPopularCategories } from './getPopularCategories';
-import { getRegex } from './getRegex';
-import { wait, Await } from './wait';
-
-export {
-  Await,
-  cn,
-  getBooks,
-  getCategories,
-  getContents,
-  getInvoices,
-  getPaths,
-  getPopularCategories,
-  getRegex,
-  wait,
-};
